requirements.txt
# Job Crawler and Clustering Pipeline Requirements

# Core dependencies
playwright==1.46.0
python-dotenv==1.0.1
psutil

# LangGraph and workflow
langgraph>=0.2.0
langchain==0.2.14
langchain-core>=0.2.32,<0.3.0

# Clustering and embeddings
sentence-transformers>=3.1.0  # For JobBERT-v2 and other embeddings
hdbscan>=0.8.0
scikit-learn>=1.3.0
numpy>=1.24.0

# LLM integration
openai>=1.0.0
tenacity>=8.0.0

# Data processing
pydantic>=2.0.0
requests>=2.31.0
pandas>=2.0.0

# Database
lancedb>=0.5.0
pyarrow>=14.0.0

# Optional but recommended
tqdm>=4.65.0
PyPDF2
rich>=13.0.0  # For CLI tools

# Flask web app
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-SocketIO>=5.3.0
python-socketio>=5.10.0
Werkzeug>=3.0.0
eventlet>=0.33.0  # For WebSocket support
python-docx>=1.0.0  # For DOCX file processing

# Testing and development
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-mock>=3.0.0
pytest-cov>=4.0.0
